remove unnecessary (?) compiler pass
